+++ src/cmd/services/m3dbnode/main/main_index_test.go
@@ -327,6 +327,7 @@ db:
         bootstrappers:
             - filesystem
             - commitlog
+            - uninitialized
         fs:
             numProcessorsPerCPU: 0.125
